Don't auto-activate outline-mode
This looks like a good idea (to me at least), but it makes it very
easy to introduce confusing bugs.  Consider the obvious configuration
line:

(add-hook 'outline-minor-mode-hook 'outshine-minor-mode)

this actually triggers a Lisp stack overflow, since
outshine-minor-mode will call outline-minor-mode, which will run its
hooks, calling back `outshine-minor-mode`, and so on.  Since hooks are
run when the mode is activated *or* deactivated, there's no simple way
to make this predictible and safe.